Framework for client update
Added ability to toggle last_skill back to -1 if you right-click the same skill again
Can now right-click an attribute to set last_skill to 50-54
Moved meta display to its own function.
Meta display can be changed based on last_skill now, allowing more potential data to be displayed without needing to slap a scroll bar on it.
Need to evaluate all the calculations needed for the information that will be displayed